Login + Rol Admin agregados

+++ routes/carts.router.js
@@ -29,20 +29,20 @@ router.post('/', uploader.single('file'), async function(req,res){
     let products = req.body
     for(let i = 0; i< products.length ; i++)
         if(! await prodDBManager.getById(products[i].id)) res.status(400).send({status:"Error", error:"Producto no Encontrado en productos.JSON"})
-    cartManager.addProducts(products)
+    await cartManager.addProducts(products)
     const result =  await cartDBManager.saveCarts()
     res.send({status:"OK", message:"Creado exitosamente",payload:result})  
 })
 
-router.post('/:cid/products/:pid',uploader.single('file'), function(req,res){
+router.post('/:cid/products/:pid',uploader.single('file'), async(req,res)=>{
     let cid = req.params.cid
     let pid = req.params.pid
-    let product = cartDBManager.getById(pid)
+    let product = await cartDBManager.getById(pid)
     if(product.length>1) res.status(400).send({status:"Error", error:"Solo se puede ingresar de a un producto en este metodo"})
     if(!product || !cid || !pid) res.status(400).send({status:"Error", error:"Producto o carrito no ingresado"})
     else{ 
         //cartManager.addProducts(cid, product, res)
-        const result = cartDBManager.saveProduct(pid,cid)
+        const result = await cartDBManager.saveProduct(pid,cid)
         
         res.send({status:"OK", message:"Producto Agregado Correctamente", payload:result})
     }
@@ -51,8 +51,12 @@ router.post('/:cid/products/:pid',uploader.single('file'), function(req,res){
 router.delete('/:cid/products/:pid', async(req,res)=>{
     let cid = req.params.cid
     let pid = req.params.pid
-    await cartDBManager.deleteProduct(cid,pid)
-    res.send({status:"OK", message:"Productos completamente eliminados"})
+    let prod = await prodDBManager.getById(pid)
+    if(prod){
+        await cartDBManager.deleteProduct(cid,pid)
+        res.send({status:"OK", message:"Productos completamente eliminados"})
+    }
+    else res.status(400).send({status:"Error", error:"No se encontró el producto"})
 })
 
 router.put('/:cid', async(req,res)=>{
@@ -61,7 +65,7 @@ router.put('/:cid', async(req,res)=>{
     if(cart._id){
         res.status(400).send({status:"Error", error:"No se puede modificar ID"})
     }
-    else cartDBManager.updateCart(cid, cart)
+    else await cartDBManager.updateCart(cid, cart)
     res.send({status:"OK", message:"Productos modificados"})
 })
 
@@ -69,17 +73,22 @@ router.put('/:cid/products/:pid', async(req,res)=>{
     let cid = req.params.cid
     let pid = req.params.pid
     let quant = req.body
-    if(quant.quantity > 0){
+    let prod = await prodDBManager.getById(pid)
+    if(quant.quantity > 0 && prod){
         quant = quant.quantity
-        cartDBManager.updateProdCart(cid,pid,quant)
+        await cartDBManager.updateProdCart(cid,pid,quant)
         res.send({status:"OK", message:"Productos modificados"})
     } 
-    else res.status(400).send({status:"Error", error:"No hay cantidad a modificar"})
+    else res.status(400).send({status:"Error", error:"Ups! Algo salió mal"})
 })
 
 router.delete('/:cid', async(req,res)=>{
     let cid = req.params.cid
-    cartDBManager.deleteAll(cid)
-    res.send({status:"OK", message:"Todos los productos han sido completamente eliminados"})
+    let cart = await cartDBManager.getById(cid)
+    if(cart){
+        await cartDBManager.deleteAll(cid)
+        res.send({status:"OK", message:"Todos los productos han sido completamente eliminados"})
+    }
+    else res.status(400).send({status:"Error", error:"No se encuentra el Carrito"})
 })
-export default router
+export default router
